Added two new logo animations and begun layout of rest of site.
The two new logo animations are blast-apart and video-mask.
The video mask doesn't have a seperate .svg-file, but does, however need a .mp4-file which is not included in the commit. It will soon be made available online so it doesn't have to be collected from the server itself.
The layout of the rest of the side now includes a bar at the bottom with a lighter background to draw the eye to the rest of the site.
There is also a link to the CSS in the nav-bar. This is temporary and only for development.